==> step_1.ipynb <==
import pandas as pd

url_v = 'https://drive.google.com/file/d/1QosQQ4RRNR9rkL4t7sB707h2Uy0XfYJe/view'
url_v = 'https://drive.google.com/uc?id=' + url_v.split('/')[-2]
visits = pd.read_csv(url_v)
visits.date = pd.to_datetime(visits.date)
#visits.head()

url_r = 'https://drive.google.com/file/d/1AeQz0kaSgz0lxYSDtuNm36muhy5fRCzZ/view'
url_r = 'https://drive.google.com/uc?id=' + url_r.split('/')[-2]
regs = pd.read_csv(url_r)
regs.date = pd.to_datetime(regs.date)
#regs.head()

visits = visits[(visits['date'] >= '2023-03-01') & (visits['date'] < '2023-09-01')]

regs = regs[(regs['date'] >= '2023-03-01') & (regs['date'] < '2023-09-01')]
#regs.head()

visits.describe()

regs.describe()
